+++ src/negmas/helpers/__init__.py
@@ -14,6 +14,8 @@ from .logging import *
 from .timeout import *
 from .misc import *
 from .numba_checks import *
+from .parallel import *
+from . import parallel
 
 __all__ = (
     types.__all__
@@ -22,6 +24,7 @@ __all__ = (
     + timeout.__all__
     + misc.__all__
     + numba_checks.__all__
+    + parallel.__all__
     # + ["prob", "inout", "numeric"]
     # + numeric.__all__
     # + prob.__all__